Fix dragging size and some refactor
